+++ feature/statistics/src/main/kotlin/dev/hossain/devicecatalog/feature/statistics/DeviceStatsPresenter.kt
@@ -6,9 +6,9 @@ import androidx.compose.runtime.produceState
 import com.slack.circuit.codegen.annotations.CircuitInject
 import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.runtime.presenter.Presenter
+import dev.hossain.devicecatalog.core.common.ExampleAppVersionService
 import dev.hossain.devicecatalog.core.data.AndroidDeviceRepository
 import dev.hossain.devicecatalog.core.data.DeviceStats
-import dev.hossain.devicecatalog.core.common.ExampleAppVersionService
 import dev.zacsweers.metro.AppScope
 import dev.zacsweers.metro.Assisted
 import dev.zacsweers.metro.AssistedFactory